\name{hypertension}
\alias{hypertension}
\docType{data}
\title{
 Diameters of veins and arteries in diabetic cases.
}
\description{
   Mean diameters of veins and arteries measured by four ophthalmologists
  at the intersections between the vessels and three concentric
  circumferences. The first circumference (with radius R) approximates
  the optic disk. The second and third circumferences are concentric
  with the first one and have radii 2R and 3R respectively.   
}
\usage{data(hypertension)}
\format{
  The format is:
 chr "hypertension"
}
\details{
  A complete description of the dataset can be found in
  
}
\source{
  G. Ayala, M.C. Lopez-Diaz, M. Lopez-Diaz and
  L. Martinez-Costa.  Methods and algorithms to test the simplex
  and Hausdorff dispersion orders with a simulation study and an
  Ophthalmological application. Technical Report. 2012
}
\examples{
data(hypertension)
}
\keyword{datasets, hypertension}
